Moved vessel activity selectors out of slice
