+++ applications/vessel-history/src/features/profile/components/activity/Activity.tsx
@@ -5,8 +5,10 @@ import { VariableSizeList as List } from 'react-window'
 import { Modal, Spinner } from '@globalfishingwatch/ui-components'
 import { selectResourcesLoading } from 'features/resources/resources.slice'
 import { VesselWithHistory } from 'types'
-import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
-import { selectFilteredEvents } from 'features/vessels/activity/vessels-activity.selectors'
+import {
+  RenderedEvent,
+  selectFilteredEvents,
+} from 'features/vessels/activity/vessels-activity.selectors'
 import { fetchRegionsThunk } from 'features/regions/regions.slice'
 import ActivityFilters from 'features/profile/filters/ActivityFilters'
 import ActivityItem from './ActivityItem'
