+++ applications/vessel-history/src/features/profile/components/activity/ActivityDate.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react'
 import { DateTime } from 'luxon'
 import I18nDate from 'features/i18n/i18nDate'
-import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
+import { RenderedEvent } from 'features/vessels/activity/vessels-activity.selectors'
 import styles from './Activity.module.css'
 
 interface ActivityDateProps {
