+++ applications/vessel-history/src/features/profile/components/activity/ActivityItem.tsx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react'
 import cx from 'classnames'
 import { Icon, IconButton } from '@globalfishingwatch/ui-components'
-import { RenderedEvent } from 'features/vessels/activity/vessels-activity.slice'
+import { RenderedEvent } from 'features/vessels/activity/vessels-activity.selectors'
 import { getEncounterStatus } from 'features/vessels/activity/vessels-activity.utils'
 import ActivityDate from './ActivityDate'
 import styles from './Activity.module.css'
